+++ setup.cfg
@@ -12,7 +12,7 @@ classifiers =
     License :: CC0 1.0 Universal (CC0 1.0) Public Domain Dedication
 
 [options]
-packages = ethereum, ethereum/vm
+packages = ethereum, ethereum/vm, ethereum/vm/instructions
 package_dir =
     =src
