+++ src/ethereum/vm/instructions/computation.py
@@ -1,0 +1,27 @@
+"""
+Ethereum Virtual Machine (EVM) Computation Instructions
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+.. contents:: Table of Contents
+    :backlinks: none
+    :local:
+
+Introduction
+------------
+
+Implementations of the EVM Arithmetic instructions.
+"""
+
+from .. import Evm
+
+
+def stop(evm: Evm) -> None:
+    """
+    Stop further execution of EVM code.
+
+    Parameters
+    ----------
+    evm :
+        The current EVM frame.
+    """
+    evm.running = False
