+++ src/ethereum/vm/interpreter.py
@@ -17,7 +17,7 @@ from typing import List, Tuple
 from ..base_types import U256, Uint
 from ..eth_types import Address, Log
 from . import Environment, Evm
-from .ops import op_implementation
+from .instructions import Ops, op_implementation
 
 
 def process_call(
@@ -85,7 +85,7 @@ def process_call(
         evm.env.state[evm.current].balance += evm.value
 
     while evm.running:
-        op = evm.code[evm.pc]
+        op = Ops(evm.code[evm.pc])
         op_implementation[op](evm)
         evm.pc += 1
